refactor(api): move remove-authentication-method controller
in identity-access-management
